refactor: 직관 기록 관련 DTO 및 모델 이름 변경
- CheckInHistoryDto -> CheckInGameDto
- CheckInHistoriesResponse -> CheckInHistoryResponse
- TeamDto -> CheckInGameTeamDto
- TeamItem -> CheckInGameTeamItem
