+++ android/app/src/main/java/com/yagubogu/data/datasource/CheckInsRemoteDataSource.kt
@@ -4,7 +4,7 @@ import com.yagubogu.data.dto.request.CheckInRequest
 import com.yagubogu.data.dto.response.CheckInCountsResponse
 import com.yagubogu.data.dto.response.FanRateResponse
 import com.yagubogu.data.dto.response.VictoryFairyRankingResponse
-import com.yagubogu.data.dto.response.checkIns.CheckInHistoriesResponse
+import com.yagubogu.data.dto.response.checkin.CheckInHistoryResponse
 import com.yagubogu.data.service.CheckInsApiService
 import com.yagubogu.data.util.safeApiCall
 import java.time.LocalDate
@@ -45,7 +45,7 @@ class CheckInsRemoteDataSource(
     override suspend fun getCheckInHistories(
         year: Int,
         filter: String,
-    ): Result<CheckInHistoriesResponse> =
+    ): Result<CheckInHistoryResponse> =
         safeApiCall {
             checkInsApiService.getCheckInHistories(year, filter)
         }
