--- android/app/src/main/java/com/yagubogu/data/dto/response/checkIns/CheckInHistoriesResponse.kt
@@ -1,10 +1,0 @@
-package com.yagubogu.data.dto.response.checkIns
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CheckInHistoriesResponse(
-    @SerialName("checkInHistory")
-    val checkInHistories: List<CheckInHistoryDto>,
-)
