+++ android/app/src/main/java/com/yagubogu/data/dto/response/checkin/CheckInGameDto.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.data.dto.response.checkIns
+package com.yagubogu.data.dto.response.checkin
 
 import com.yagubogu.presentation.stats.attendance.AttendanceHistoryItem
 import kotlinx.serialization.SerialName
@@ -6,15 +6,15 @@ import kotlinx.serialization.Serializable
 import java.time.LocalDate
 
 @Serializable
-data class CheckInHistoryDto(
+data class CheckInGameDto(
     @SerialName("checkInId")
     val checkInId: Int, // 직관 ID
     @SerialName("stadiumFullName")
     val stadiumFullName: String, // 경기장 전체 이름
     @SerialName("homeTeam")
-    val homeTeam: TeamDto,
+    val homeTeam: CheckInGameTeamDto,
     @SerialName("awayTeam")
-    val awayTeam: TeamDto,
+    val awayTeam: CheckInGameTeamDto,
     @SerialName("attendanceDate")
     val attendanceDate: String, // 직관 날짜 (예시 "2025-04-05")
 ) {
