+++ android/app/src/main/java/com/yagubogu/data/dto/response/checkin/CheckInHistoryResponse.kt
@@ -1,0 +1,10 @@
+package com.yagubogu.data.dto.response.checkin
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CheckInHistoryResponse(
+    @SerialName("checkInHistory")
+    val checkInHistory: List<CheckInGameDto>,
+)
